refactor: move auth pages
