+++ pages/auth/EmailVerificationPage.tsx
